pyproject.toml
[project]
name = "ml-intern"
version = "0.2.0"
description = "An autonomous ML engineering agent for the Hugging Face ecosystem, built on the Claude Agent SDK."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # Core dependencies
    "datasets>=4.4.1",
    "pydantic>=2.12.3",
    "python-dotenv>=1.2.1",
    # Agent runtime (Claude Agent SDK replaces litellm + custom loop)
    "claude-agent-sdk>=0.1.0",
    "anthropic>=0.40.0",
    "requests>=2.33.0",
    "huggingface-hub>=1.0.1",
    "prompt-toolkit>=3.0.0",
    "thefuzz>=0.22.1",
    "rich>=13.0.0",
    "nbconvert>=7.16.6",
    "nbformat>=5.10.4",
    "whoosh>=2.7.4",
    "beautifulsoup4>=4.12.0",
    # Web backend dependencies
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "httpx>=0.27.0",
    "websockets>=13.0",
]

[project.optional-dependencies]

# Evaluation/benchmarking dependencies
eval = [
    "inspect-ai>=0.3.149",
    "pandas>=2.3.3",
    "datasets>=4.3.0",
    "tenacity>=8.0.0",
]

# Development and testing dependencies
dev = [
    "pytest>=9.0.2",
]

# All dependencies (eval + dev)
all = [
    "hf-agent[eval,dev]",
]

[project.scripts]
ml-intern = "agent.main:cli"

[build-system]
requires = ["setuptools>=64"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["agent*"]

[tool.uv]
package = true
